# vtol_att_control: Add parameter for VTOL PWM value in FW mode